pyproject.toml
[project]
name = "disaster-sms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "openai>=1.93.1",
    "paho-mqtt>=2.1.0",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "twilio>=9.6.4",
]
